Plane Cont'd
-And yet again I'm fiddling with aviary mechanics. There is no other way
around than to implement the third vector. That is, velRotation (and
this.rotation with it) is the first and cursorRotation is the second.
The third vector (could be this.rotation) is going to take part of
velRotation's responsibility. The idea is to make this.rotation =
cursorRotation (possibly, required for shoot aiming) and leave
velRotation responsible only for the direction of plane motion.
Eventually, this system will also be replaced by forces which inertia
also counts.