Plane Cont'd
-Now plane does not switch sides after landing. This is possible for the
following.
-Added boolean property 'landed' to the plane prototype. Landed will be
used in the future more as options upon landing will become available.
So far, 'landed' is used to keep the plane from switching sides.
-Added boolean property 'landedLeft' to the plane prorotype.
'landedLeft' is used to keep track of which side the plane landed on and
make it face to that side (by setting this.velRotation to either 0 or
180.)
+Plane always takes off to the right. Fix it.
+Redo the calculations on the line 58. this.cursorRotation shouldnot be
there.
+Follow the plan.