Bullets Cont'd: Broken Damage Fixed
-Oh, by the way. The issue with the plane harming itself happened
because bullets were not hitting the body of the plane but the
horizontal and vertical speed pars of plane.test. The bullets were
hitting and damaging the UI. And since plane.test is also in
plane.children[] (perhaps the reason, maybe the link is one of other
ones,) plane.hitTest() might have considered plane.test a part of plane.