Bullets Concluded. Score UI added
-Numerous changes made all over (settings, UI, even Bullet constructor
parameters) to accomodate for the score object attached to plane and
properly updated upon certain events (death, kill, etc.)
-Bullet now has a function freeze(). Incidentally, bullet.frozen boolean
property is used to keep track whether the bullet must follow the plane
or not. Combined, they allow the bullet to stop its relative motion to
the plane. The function is designed for test purposes only.
-settings has new objects: settigs.UI, settings.score, and
settings.score.multipliers.
-Added points, deaths, and kills createjs.Text objects to UI class;
-Added UI.setPoints(), UI.setDeaths(), UI.setKills() to operate the
newly added properties.
+Add an object representing the other planes.
+Ojbect's properties: x, y, velX, velY, ... look into plane.js
properties for the rest.
+Add hitTest() handling for the othe planes.